Update the Twitter Fetcher component's custom JavaScript.
This change modernizes the custom JavaScript we use with Twitter Post Fetcher. Before, we were using JS compatible with IE11. Now that we don't support IE11, we can use things like template literals, `Intl`, and multiple classes in a `classList.add` method.

This change should make it easier for those wanting to customize the Twitter template to more easily be able to do so. It's also faster than the previous implementation by 200ms-1.2s. We're also calling this custom JS in a separate file that the browser should cache after the first visit.

This change also improves the accessibility of this component by implementing aria-label instead of aria-labelledby for the "Posted X hours ago" permalinks.